Working basis for sigsys-based sandbox.
This is not very performant (lots of context switches) and requires a lot of
error-prone code in order to correctly marshal all of the syscalls (dear god
the switch statements). As a result, this sandbox type will not be further
developed and remains merely as a reminder of how to go about doing it.

A future commit will transition this to a sandbox based on LD_PRELOAD and
overriding functions in glibc, which allows for directly marshalling things
to and from the parent rather than needing the kernel to raise a signal first.