Kill preload_dir with fire.
I figured out a sane way to implement mmap on fds so that .so files can be loaded
in even after the sandbox has started. This required giving a lot of access to the
filesystem (namely every .so file in /lib and /lib64), but the parent can futher
restrict that if necessary.